refactor: reorganize reference components
