+++ packages/ui/core-components/src/lib/unsorted/viz/index.js
@@ -36,8 +36,7 @@ export { default as SankeyChart } from './sankey/SankeyChart.svelte';
 export { default as ScatterPlot } from './scatter/ScatterPlot.svelte';
 export { default as Scatter } from './scatter/Scatter.svelte';
 export { default as SearchBar } from './core/SearchBar.svelte';
-export { default as ReferenceLine } from './references/ReferenceLine.svelte';
-export { default as ReferenceArea } from './references/ReferenceArea.svelte';
+export * from './references';
 export { default as USMap } from './map/USMap.svelte';
 export { default as Value } from './core/Value.svelte';
 export { default as ValueError } from './core/ValueError.svelte';
