+++ packages/ui/core-components/src/lib/unsorted/viz/references/ReferenceArea/ReferenceArea.svelte
@@ -3,10 +3,10 @@
 </script>
 
 <script>
-	import { QueryLoad } from '../../../atoms/query-load';
+	import { QueryLoad } from '../../../../atoms/query-load';
 	import ReferenceArea from './_ReferenceArea.svelte';
-	import EmptyChart from '../core/EmptyChart.svelte';
-	import ErrorChart from '../core/ErrorChart.svelte';
+	import EmptyChart from '../../core/EmptyChart.svelte';
+	import ErrorChart from '../../core/ErrorChart.svelte';
 	import { Query } from '@evidence-dev/sdk/usql';
 
 	export let data = undefined;
