+++ packages/ui/core-components/src/lib/unsorted/viz/references/ReferenceArea/index.js
@@ -1,0 +1,1 @@
+export { default as ReferenceArea } from './ReferenceArea.svelte';
