+++ packages/ui/core-components/src/lib/unsorted/viz/references/index.js
@@ -1,0 +1,2 @@
+export * from './ReferenceArea';
+export * from './ReferenceLine';
